refactor: getApplications 메서드를 도메인 구조에 맞게 패키지 이동 [#back-31]
